changes fix dropdown issue #23 navbar #22 and #21
For #22, I delete 'my cart' because I don't think we need this function for our site. For #23 and #21, now in homepage for those not registered people, if they click 'sell stuff/your store/log in', page will jump to login page, and if they didn't registered, they can click on 'join us' to create an account in register page. I think it is important that there are "sell stuff/your store" button in home page also for not-registered people so they will know what this website actually is. And when they click these buttons, they may create an account. The change is not lot, if it is ok for everyone, I will merge this.